logcapture: actually redirect output
Turns out that captureLog operated on the default (global) logrus logger
and never actually redirected output to t.Logf.  Fix this, here, and
also add the logger to the output so that we can redirect the output
back (even though it's not obvious to me why we need to do so).

There is still some output from the observer (which has its own logrus
instance). See subsequent commits.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>